Move main code into trait
This allows SubStringy to be used as a mixin within a subclass of
Stringy which is combining multiple extensions.
